Fix #3
Names / ids of the elements in the opal popup were different to those in exam. Now the user gets redirected to another login page for opal, where the names are the same. Also the index of the TU Dresden option is now searched by using the text instead of hardcoded values because it differs between opal and exam.